Module 7 task 1
correcting module form
